Fix that overridden event.preventDefault() did not call the original
